Fix mysensors unload clean up (#47541)
